refactoring for sparse representation

+++ src/main/scala/InstanceWiseAbstractScore.scala
@@ -4,7 +4,7 @@ abstract class InstanceWiseAbstractScore extends Serializable {
     
     def getResult(matWithClass: Array[Array[Long]],
                   matWithFeatures: Array[Array[Array[Long]]],
-                  selectedVariablesArray: Array[Int],
+                  selectedVariablesArray: Array[Long],
                   variableLevels: Array[Double],
                   classLevels: Array[Double]): Double
